requirements.txt
flask
flask_talisman
google-cloud-firestore
python-dotenv
requests
flask-cors
PyJWT
werkzeug
gunicorn